finalize gnn-assisted-sampling code
