+++ pySTK-master/code/common.py
@@ -3,5 +3,5 @@ enum_dict = {
     'rand': 0, 'hist': 1, 'hist_grad': 2, 'hist_grad_rand': 3, 'grad': 4, 'lcc': 5, 'max': 6, 'similarity': 7,
     'max_neighbor': 8, 'random_walk': 9, 'value_weighted_random_walk': 10, 'similarity_weighted_random_walk': 11,
     'max_seeded_random_walk': 12, 'max_seeded_weighted_random_walk': 13, 'complex_max_seeded_random_walk': 14,
-    'hist_lcc': 15, 'grad_lcc': 16, 'lcc_rand': 17, 'lcc_skip': 18
+    'hist_lcc': 15, 'grad_lcc': 16, 'gnn': 17
     }
