+++ pySTK-master/code/setup.sh
