--- pySTK-master/scripts/ENV.yml
@@ -1,189 +1,0 @@
-name: intellSampler
-channels:
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=5.1=1_gnu
-  - argon2-cffi=21.3.0=pyhd3eb1b0_0
-  - argon2-cffi-bindings=21.2.0=py37h7f8727e_0
-  - attrs=21.4.0=pyhd3eb1b0_0
-  - backcall=0.2.0=pyhd3eb1b0_0
-  - beautifulsoup4=4.11.1=py37h06a4308_0
-  - blas=1.0=openblas
-  - bleach=4.1.0=pyhd3eb1b0_0
-  - bottleneck=1.3.4=py37hce1f21e_0
-  - brotli=1.0.9=he6710b0_2
-  - bzip2=1.0.8=h7b6447c_0
-  - c-ares=1.18.1=h7f8727e_0
-  - ca-certificates=2022.4.26=h06a4308_0
-  - cached-property=1.5.2=py_0
-  - certifi=2022.6.15=py37h06a4308_0
-  - cffi=1.15.0=py37hd667e15_1
-  - curl=7.82.0=h7f8727e_0
-  - cycler=0.11.0=pyhd3eb1b0_0
-  - dbus=1.13.18=hb2f20db_0
-  - debugpy=1.5.1=py37h295c915_0
-  - decorator=5.1.1=pyhd3eb1b0_0
-  - defusedxml=0.7.1=pyhd3eb1b0_0
-  - entrypoints=0.4=py37h06a4308_0
-  - expat=2.4.4=h295c915_0
-  - fontconfig=2.13.1=h6c09931_0
-  - freetype=2.11.0=h70c0345_0
-  - future=0.18.2=py37_1
-  - giflib=5.2.1=h7b6447c_0
-  - glib=2.69.1=h4ff587b_1
-  - gst-plugins-base=1.14.0=h8213a91_2
-  - gstreamer=1.14.0=h28cd5cc_2
-  - h5py=2.10.0=py37h7918eee_0
-  - hdf4=4.2.13=h3ca952b_2
-  - hdf5=1.10.4=hb1b8bf9_0
-  - icu=58.2=he6710b0_3
-  - importlib-metadata=4.11.3=py37h06a4308_0
-  - importlib_metadata=4.11.3=hd3eb1b0_0
-  - importlib_resources=5.2.0=pyhd3eb1b0_1
-  - ipykernel=6.9.1=py37h06a4308_0
-  - ipython=7.31.1=py37h06a4308_0
-  - ipython_genutils=0.2.0=pyhd3eb1b0_1
-  - ipywidgets=7.6.5=pyhd3eb1b0_1
-  - jedi=0.18.1=py37h06a4308_1
-  - jinja2=3.0.3=pyhd3eb1b0_0
-  - joblib=1.1.0=pyhd3eb1b0_0
-  - jpeg=9e=h7f8727e_0
-  - jsoncpp=1.9.4=hff7bd54_2
-  - jsonschema=4.4.0=py37h06a4308_0
-  - jupyter=1.0.0=py37_7
-  - jupyter_client=7.2.2=py37h06a4308_0
-  - jupyter_console=6.4.3=pyhd3eb1b0_0
-  - jupyter_core=4.10.0=py37h06a4308_0
-  - jupyterlab_pygments=0.1.2=py_0
-  - jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
-  - krb5=1.19.2=hac12032_0
-  - lcms2=2.12=h3be6417_0
-  - ld_impl_linux-64=2.38=h1181459_1
-  - libcurl=7.82.0=h0b77cf5_0
-  - libedit=3.1.20210910=h7f8727e_0
-  - libev=4.33=h7f8727e_1
-  - libffi=3.3=he6710b0_2
-  - libgcc-ng=11.2.0=h1234567_1
-  - libgfortran-ng=7.5.0=ha8ba4b0_17
-  - libgfortran4=7.5.0=ha8ba4b0_17
-  - libgomp=11.2.0=h1234567_1
-  - libnetcdf=4.6.1=h11d0813_2
-  - libnghttp2=1.46.0=hce63b2e_0
-  - libogg=1.3.5=h27cfd23_1
-  - libopenblas=0.3.18=hf726d26_0
-  - libpng=1.6.37=hbc83047_0
-  - libsodium=1.0.18=h7b6447c_0
-  - libssh2=1.10.0=h8f2d780_0
-  - libstdcxx-ng=11.2.0=h1234567_1
-  - libtheora=1.1.1=h7f8727e_3
-  - libtiff=4.1.0=h2733197_0
-  - libuuid=1.0.3=h7f8727e_2
-  - libvorbis=1.3.7=h7b6447c_0
-  - libwebp=1.2.2=h55f646e_0
-  - libwebp-base=1.2.2=h7f8727e_0
-  - libxcb=1.15=h7f8727e_0
-  - libxml2=2.9.12=h03d6c58_0
-  - lz4-c=1.8.1.2=h14c3975_0
-  - markupsafe=2.1.1=py37h7f8727e_0
-  - matplotlib-base=3.5.1=py37ha18d171_1
-  - matplotlib-inline=0.1.2=pyhd3eb1b0_2
-  - mistune=0.8.4=py37h14c3975_1001
-  - munkres=1.1.4=py_0
-  - nbclient=0.5.13=py37h06a4308_0
-  - nbconvert=6.4.4=py37h06a4308_0
-  - nbformat=5.3.0=py37h06a4308_0
-  - ncurses=6.3=h7f8727e_2
-  - nest-asyncio=1.5.5=py37h06a4308_0
-  - notebook=6.4.11=py37h06a4308_0
-  - numexpr=2.8.1=py37hecfb737_0
-  - numpy=1.21.5=py37h7a5d4dd_2
-  - numpy-base=1.21.5=py37hb8be1f0_2
-  - openssl=1.1.1q=h7f8727e_0
-  - packaging=21.3=pyhd3eb1b0_0
-  - pandas=1.3.5=py37h8c16a72_0
-  - pandocfilters=1.5.0=pyhd3eb1b0_0
-  - parso=0.8.3=pyhd3eb1b0_0
-  - pcre=8.45=h295c915_0
-  - pexpect=4.8.0=pyhd3eb1b0_3
-  - pickleshare=0.7.5=pyhd3eb1b0_1003
-  - pip=21.0.1=py37h06a4308_0
-  - prometheus_client=0.13.1=pyhd3eb1b0_0
-  - prompt-toolkit=3.0.20=pyhd3eb1b0_0
-  - prompt_toolkit=3.0.20=hd3eb1b0_0
-  - ptyprocess=0.7.0=pyhd3eb1b0_2
-  - pycparser=2.21=pyhd3eb1b0_0
-  - pygments=2.11.2=pyhd3eb1b0_0
-  - pyqt=5.9.2=py37h05f1152_2
-  - pyrsistent=0.18.0=py37heee7806_0
-  - python=3.7.11=h12debd9_0
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - python-fastjsonschema=2.15.1=pyhd3eb1b0_0
-  - pytz=2022.1=py37h06a4308_0
-  - pyzmq=22.3.0=py37h295c915_2
-  - qt=5.9.7=h5867ecd_1
-  - qtconsole=5.3.0=pyhd3eb1b0_0
-  - qtpy=2.0.1=pyhd3eb1b0_0
-  - readline=8.1.2=h7f8727e_1
-  - scikit-learn=1.0.2=py37h51133e4_1
-  - scipy=1.7.3=py37h492baa0_0
-  - seaborn=0.11.2=pyhd3eb1b0_0
-  - send2trash=1.8.0=pyhd3eb1b0_1
-  - setuptools=61.2.0=py37h06a4308_0
-  - sip=4.19.8=py37hf484d3e_0
-  - six=1.16.0=pyhd3eb1b0_0
-  - soupsieve=2.3.1=pyhd3eb1b0_0
-  - sqlite=3.36.0=hc218d9a_0
-  - tbb=2021.5.0=hd09550d_0
-  - termcolor=1.1.0=py37h06a4308_1
-  - terminado=0.13.1=py37h06a4308_0
-  - testpath=0.6.0=py37h06a4308_0
-  - threadpoolctl=2.2.0=pyh0d69192_0
-  - tk=8.6.11=h1ccaba5_0
-  - tornado=6.1=py37h27cfd23_0
-  - tqdm=4.64.0=py37h06a4308_0
-  - traitlets=5.1.1=pyhd3eb1b0_0
-  - typing_extensions=4.1.1=pyh06a4308_0
-  - wcwidth=0.2.5=pyhd3eb1b0_0
-  - webencodings=0.5.1=py37_1
-  - wheel=0.37.0=pyhd3eb1b0_1
-  - widgetsnbextension=3.5.2=py37h06a4308_0
-  - xz=5.2.5=h7f8727e_1
-  - zeromq=4.3.4=h2531618_0
-  - zipp=3.8.0=py37h06a4308_0
-  - zlib=1.2.12=h7f8727e_2
-  - zstd=1.3.7=h0b5b093_0
-  - pip:
-    - aiohttp==3.8.1
-    - aiosignal==1.2.0
-    - async-timeout==4.0.2
-    - asynctest==0.13.0
-    - autobahn==22.5.1
-    - automat==20.2.0
-    - charset-normalizer==2.1.0
-    - constantly==15.1.0
-    - cryptography==37.0.3
-    - fonttools==4.33.3
-    - frozenlist==1.3.0
-    - hyperlink==21.0.0
-    - idna==3.3
-    - incremental==21.3.0
-    - kiwisolver==1.4.3
-    - llvmlite==0.38.1
-    - matplotlib==3.5.2
-    - multidict==6.0.2
-    - naturalneighbor==0.2.1
-    - numba==0.55.2
-    - param==1.12.2
-    - pillow==9.1.1
-    - pymp-pypi==0.5.0
-    - pyparsing==3.0.9
-    - scikit-network==0.26.0
-    - twisted==22.4.0
-    - txaio==22.2.1
-    - typing-extensions==4.2.0
-    - vtk==9.0.3
-    - wslink==1.6.5
-    - yarl==1.7.2
-    - zope-interface==5.4.0
-prefix: /vast/home/wanyef/.conda/envs/intellSampler
